Add support for Rails 7
Why these changes are being introduced:

This gem currently supports Rails 5 and 6 only, but we are starting
to write applications that use Rails 7.

Relevant ticket(s):

https://github.com/MITLibraries/timdex-ui/issues/16

How this addresses that need:

This updates the gemspec to support versions of Rails greater than
or equal to 5 but less than 8.

Side effects of this change:

* This gem still uses sassc, which appears to be unsupported. We
should investigate whether it's possible to use a more modern
CSS preprocessor (see #24).
* Further work may be needed for the this gem to function
alongside advanced Rails 7 UI features like Hotwire and Turbolinks
(see #25).