Allow for optional platform names in header
Why these changes are being introduced:

Matt added the platform name to Omeka to remind users where in
the MITL ecosystem they are. We were planning to add this to
Geodata when we realized that it would probably be useful in other
Rails apps, and should thus go in the theme gem.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/GDT-124

How this addresses that need:

This adds an optional platform name in the header, based on the
`PLATFORM_NAME` environment variable. Environments that don't set
`PLATFORM_NAME` will continue to have the standard header.

Side effects of this change:

The SCSS updates for this should likely be moved upstream to the
style guide, but I couldn't decide if that should happen before or
after this merges.