Display title in nav only if PLATFORM_NAME is not set
Why these changes are being introduced:

The app name is inserted into the header based on the `PLATFORM_NAME`
environment variable. If this variable is set, the title should
not display

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/GDT-124

How this addresses that need:

This makes the title in the nav render if `PLATFORM_NAME` is not
set.

Side effects of this change:

We will need to make this change downstream, as most (if not all)
of our Rails properties have customized the site nav partial.