Add integrity hashes for remote scripts and stylesheets
Why these changes are being introduced:

It's good practice to validate checksums for external libraries.
We've done this in individual apps, but it makes sense to make
the change in the theme gem.

Relevant ticket(s):

* [ENGX-290](https://mitlibraries.atlassian.net/browse/ENGX-290)

How this addresses that need:

This adds `integrity`, `crossorigin`, and `referrerpolicy`
attributes for remotely hosted scripts and stylesheets.

Side effects of this change:

I'm not totally sure how to exhaustively test these changes. I've
confirmed that the scripts/stylesheets load as expected, and
checked the changes locally in a few of our apps. That feels like
enough, but just signaling the uncertainty here in case the
reviewer has additional insight.